# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "2.0.0"
  title: pl-song-service
basePath: /v1
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  api_key:
    type: apiKey
    description: Authorization token used in query parameter
    in: query
    name: api_key
  jwt_key:
    type: apiKey
    description: Authorization json web token used in header
    in: header
    name: Authorization
paths:
  /pingAllServices:
    x-swagger-router-controller: ping
    get:
      security:
        - api_key: []
      tags:
        - ping
      description: |
        Make sure authentication is working properly
      operationId: pingAllServices
      responses:
        200:
          description: Success
  /ping:
    x-swagger-router-controller: ping
    get:
      security:
        - api_key: []
      tags:
        - ping
      description: |
        Make sure authentication is working properly
      operationId: ping
      responses:
        200:
          description: Success

  /songs/batch:
    x-swagger-router-controller: songs
    get:
      security:
        - api_key: []
      tags:
        - songs
      description: |
        Gets a batch of songs from their ids.
      operationId: getBatch
      parameters:
        - name: songIDs
          in: query
          description: The ID of the song
          required: true
          type: array
          items:
            type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/AudioReplacementRequest"
        default:
          description: Error
          schema:
            $ref: "./deps/swagger-defs/commonDefinitions.yaml#/ErrorResponse"

  /songs/{songID}/completeAudioReplacement:
    x-swagger-router-controller: songs
    post:
      security:
        - api_key: []
      tags:
        - songs
      description: |
        This endpoint is the webhook used by the audio converter service to
        report that acquisition of the audio file has completed.  If successful,
        the song's 'key' property will be changed.
      operationId: completeAudioReplacement
      parameters:
        - name: songID
          in: path
          description: The ID of the song
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/AudioReplacementRequest"
        default:
          description: Error
          schema:
            $ref: "./deps/swagger-defs/commonDefinitions.yaml#/ErrorResponse"

  /songs/{songID}/requestAudioReplacement:
    x-swagger-router-controller: songs
    post:
      security:
        - api_key: []
      tags:
      - songs
      description: |
        This endpoint requests a replacement for a song's audio file.
      operationId: requestAudioReplacement
      parameters:
        - name: songID
          in: path
          description: The ID of the song
          required: true
          type: string
        - name: data
          in: body
          description: The info for replacing the song audio
          required: true
          schema:
            type: object
            required:
              -   youTubeID
            properties:
              youTubeID:
                type: string
                description: the id of the youtube video containing the desired replacement audio. 
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/AudioReplacementRequest"
        default:
          description: Error
          schema:
            $ref: "./deps/swagger-defs/commonDefinitions.yaml#/ErrorResponse"
  /songRequests/search:
    x-swagger-router-controller: songRequests
    get:
      security:
        - api_key: []
      tags:
        - songRequests
      description: |
        Search for a song request matching the searchText
      operationId: search
      parameters:
        - name: searchText
          in: query
          description: the text to search for
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "./deps/swagger-defs/commonDefinitions.yaml#/Songs"
        default:
          description: Error
          schema:
            $ref: "./deps/swagger-defs/commonDefinitions.yaml#/ErrorResponse"
  /songRequests/complete/{spotifyID}:
    x-swagger-router-controller: songRequests
    post:
      tags:
        - songRequests
      description: |
        Used to report that audio acquisition has completed -- used as a webhook
        for the audio converter
      operationId: completeSongAcquisition
      parameters:
        - name: spotifyID
          in: path
          description: The spotify_id of the acquired audio
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            type: object
            required:
              - bucketName
              - key
            properties:
              bucketName:
                type: string
                description: the name of the bucket containing the completed song
                example: playolasongstestbucket
              key:
                type: string
                description: the key of the newly stored audioFile
                example: sugar.m4a
      responses:
        200:
          description: Success
          schema:
            $ref: './deps/swagger-defs/commonDefinitions.yaml#/SongRequest'
        default:
          description: Error
          schema:
            $ref: "./deps/swagger-defs/commonDefinitions.yaml#/ErrorResponse"

  /songRequests/{spotifyID}:
    x-swagger-router-controller: songRequests
    post:
      security:
        - api_key: []
      tags:
        - songRequests
      operationId: getOrCreateSongRequest
      parameters:
        - name: spotifyID
          in: path
          description: The spotify id of the desired song
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: './deps/swagger-defs/commonDefinitions.yaml#/SongRequest'
        default:
          description: Error
          schema:
            $ref: "./deps/swagger-defs/commonDefinitions.yaml#/ErrorResponse"
    delete:
      security:
        - api_key: []
      tags:
        - songRequests
      operationId: removeMatchingSpotifyID
      parameters:
        - name: spotifyID
          in: path
          description: The spotify id of the desired song
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: './deps/swagger-defs/commonDefinitions.yaml#/SongRequest'
        default:
          description: Error
          schema:
            $ref: "./deps/swagger-defs/commonDefinitions.yaml#/ErrorResponse"

definitions:
  AudioReplacementRequest:
    type: object
    properties:
      song:
        $ref: './deps/swagger-defs/commonDefinitions.yaml#/Song'
      youtube_id:
        type: string
        example: O46hKaW5eY4
      status:
        type: string
        example: 'processing'
        description: |
          The status of the audioReplacementRequest. Possibilities are:
          * 'processing'
          * 'completed'
          * 'errored'